+++ libnodegl/backends/gl/gpu_ctx_gl.h
@@ -74,4 +74,6 @@ struct gpu_ctx_gl {
     void (*glGetQueryObjectui64v)(const struct glcontext *gl, GLuint id, GLenum pname, GLuint64 *params);
 };
 
+int ngli_gpu_ctx_gl_wrap_framebuffer(struct gpu_ctx *s, GLuint fbo);
+
 #endif